Hito3/test1BlogUnab/requirements.txt
django
djangorestframework
gunicorn
django-ckeditor